[beta 0.2] Bug fixes
Account for all of the menu items combinations in order to compute the overall income from all combinations.
The previous code was assuming 1 Customer = 1 item. But that's obviously wrong as a customer buying a 'breakfast set' is very likely to also take a coffee. Or at midday, people may take a drink and a lunch set. etc...
What the modification does not account for however, is the full tree of conditional probability. It does not for example account of the fact that if someone buys 'eggs', it is also likely to take 'bacon' and 'coffee'.
Here probabilities are daily function averages therefore.